refactor(parser): extract HTML tag set
